Poçoes Correções #85 #40 #82
Alterações:
1) mudado tipo da variavel pots de float para inteiro
2) Adicionado mais uma poção ao jogador ficando um total de 6 para o gameplay inteiro
3) mudado ouro inicial para 0

+++ jogo42-cb/jogo42.h
@@ -79,7 +79,7 @@ typedef struct // GameState
         //! Variavel usada para descrescer a saciedade
         float timerSac;
         //! Potions
-        float pots;
+        int pots;
         //! Variavel que determina se a pocao estah sendo usada
         bool usingPot;
         //! Varaivel usada para contar o tempo de ativacao da potion
